feat: lnurl-auth onboarding
