+++ src/app/screens/Enable/WebbtcEnable.tsx
@@ -29,7 +29,7 @@ export default function WebbtcEnable(props: Props) {
     }
 
     fetchAccountAndSetComponent();
-  }, [props.origin, account]);
+  }, [account]);
 
   return (
     <div>
