+++ src/extension/background-script/actions/lnurl/authOrPrompt.ts
@@ -33,10 +33,17 @@ async function authOrPrompt(
   // we have the check the unlock status manually. The account can still be locked
   // If it is locked we must show a prompt to unlock
   const isUnlocked = await state.getState().isUnlocked();
+  const account = await state.getState().getAccount();
 
   // check if there is a publisher and lnurlAuth is enabled,
   // otherwise we we prompt the user
-  if (isUnlocked && allowance && allowance.enabled && allowance.lnurlAuth) {
+  if (
+    isUnlocked &&
+    allowance &&
+    allowance.enabled &&
+    allowance.lnurlAuth &&
+    account?.mnemonic
+  ) {
     return await authFunction({ lnurlDetails, origin: message.origin });
   } else {
     try {
