--- src/extension/background-script/actions/onboard/index.ts
@@ -1,3 +1,0 @@
-import prompt from "./prompt";
-
-export { prompt };
